[skip ci] Refactor for ease of maintenance and extensibility.
