+++ include/RouteMap.h
@@ -28,11 +28,13 @@
 
 #include "ODAPI.h"
 #include "GribRecordSet.h"
+#include "ConstraintChecker.h"
+#include "RoutePoint.h"
+#include "Position.h"
+#include "Boat.h"
 
 struct RouteMapConfiguration;
 class IsoRoute;
-
-typedef std::list<IsoRoute*> IsoRouteList;
 
 class PlotData;
 
@@ -79,186 +81,6 @@ struct climatology_wind_atlas {
 };
 
 class WR_GribRecordSet;
-
-/**
- * A base class representing a geographical position that is part of a route.
- *
- * RoutePoint stores the essential information about a point along a route,
- * including its latitude and longitude coordinates, which polar was used
- * to reach this point, and how many tacks were required. It also provides
- * methods for retrieving weather data at this position and determining if
- * the route point intersects with land or boundaries.
- *
- * This serves as the foundation class for more specialized route position
- * classes like Position and PlotData.
- */
-class RoutePoint {
-public:
-  RoutePoint(double latitude = 0., double longitude = 0., int polar_idx = -1,
-             int tack_count = 0, int jibe_count = 0,
-             int sail_plan_change_count = 0, int dm = 0,
-             bool data_deficient = false)
-      : lat(latitude),
-        lon(longitude),
-        polar(polar_idx),
-        tacks(tack_count),
-        jibes(jibe_count),
-        sail_plan_changes(sail_plan_change_count),
-        grib_is_data_deficient(data_deficient),
-        data_mask(dm) {}
-
-  virtual ~RoutePoint() {};
-
-  double lat;
-  double lon;
-  /** The index of the polar to use at this position. */
-  int polar;
-  /** The cumulative number of tack maneuvers to get to this position. */
-  int tacks;
-  /** The cumulative number of jibe maneuvers to get to this position. */
-  int jibes;
-  /** The cumulative number of sail plan changes to get to this position. */
-  int sail_plan_changes;
-
-  bool grib_is_data_deficient;
-
-  bool GetPlotData(RoutePoint* next, double dt,
-                   RouteMapConfiguration& configuration, PlotData& data);
-  // Return the wind data at the route point.
-  bool GetWindData(RouteMapConfiguration& configuration, double& W, double& VW,
-                   int& data_mask);
-  // Return the current data at the route point.
-  bool GetCurrentData(RouteMapConfiguration& configuration, double& C,
-                      double& VC, int& data_mask);
-
-  // Return true if the route point crosses land.
-  bool CrossesLand(double dlat, double dlon);
-  // Return true if the route point enters a boundary.
-  bool EntersBoundary(double dlat, double dlon);
-  /**
-   * Propagates from current position to a specific target point.
-   *
-   * Attempts to find a valid route from the current position to the specified
-   * lat/lon coordinates within the constraints of the configuration.
-   *
-   * @param dlat Target latitude
-   * @param dlon Target longitude
-   * @param configuration Route configuration parameters
-   * @param heading [out] Final heading to reach target
-   * @param data_mask [out] Mask indicating data sources used
-   * @param end If true, indicates this is the final destination point
-   *
-   * @return Time in seconds to reach target point, or NAN if not possible
-   */
-  double PropagateToPoint(double dlat, double dlon, RouteMapConfiguration& cf,
-                          double& heading, int& data_mask, bool end = true);
-
-  /**
-   * Calculates boat performance based on environmental conditions and desired
-   * heading.
-   *
-   * Given a specific boat heading relative to wind, this function calculates:
-   * 1. How fast the boat will move through water (speed through water)
-   * 2. The actual direction the boat will travel through water (course through
-   * water)
-   * 3. How fast the boat will move over ground after accounting for currents
-   * (speed over ground)
-   * 4. The actual direction the boat will travel over ground (course over
-   * ground)
-   *
-   * @param configuration Boat and route configuration settings
-   * @param timeseconds Duration in seconds for the calculation
-   * @param twd [in] Wind direction over ground (degrees)
-   * @param tws [in] Wind speed over ground (knots)
-   * @param windDirOverWater [in] Wind direction through water (degrees)
-   * @param windSpeedOverWater [in] Wind speed through water (knots)
-   * @param currentDir [in] Current direction (degrees)
-   * @param currentSpeed [in] Current speed (knots)
-   * @param twa [in] True Wind Angle (TWA) (degrees)
-   * @param atlas [in] Wind climatology atlas containing probability
-   * distributions
-   * @param data_mask [in/out] Bit mask indicating data sources (GRIB,
-   * climatology, etc.)
-   * @param ctw [in] Boat's bearing relative to true wind (W+twa), initialized
-   * by caller
-   * @param stw [out] Boat speed through water
-   * @param cog [out] Boat bearing over ground
-   * @param sog [out]Boat speed over ground
-   * @param dist [out] Distance traveled over ground (nm)
-   * @param newpolar [in] Index of the polar to use from the boat's polar array
-   * @param bound [in] If true, returns NAN when wind speed is outside the range
-   * defined in the polar data. If false, extrapolates the boat speed when wind
-   * speed is outside the polar data range.
-   *
-   * @return true if computation successful, false if NaN values detected
-   */
-  bool ComputeBoatSpeed(RouteMapConfiguration& configuration,
-                        double timeseconds, double twd, double tws,
-                        double windDirOverWater, double windSpeedOverWater,
-                        double currentDir, double currentSpeed, double twa,
-                        climatology_wind_atlas& atlas, double ctw, double& stw,
-                        double& cog, double& sog, double& dist, int newpolar,
-                        bool bound = true, const char* caller = "unknown");
-
-  /**
-   * Bit flags indicating what data sources were used for wind and current
-   * calculations and other routing conditions.
-   *
-   * These flags track various aspects of each position in the routing
-   * calculation:
-   * 1. The origin of wind and current data (GRIB or climatology)
-   * 2. Whether the data was "deficient" (outside optimal time/location range)
-   * 3. Environmental conditions like day/night status
-   *
-   * The flags can be combined using bitwise OR operations to represent multiple
-   * conditions simultaneously. For example, a position at night using GRIB
-   * current data and climatology wind data would have a data_mask containing:
-   * (GRIB_CURRENT | CLIMATOLOGY_WIND | NIGHT_TIME)
-   *
-   * These flags serve multiple purposes:
-   * - Visual differentiation in the route display (different colors for data
-   * sources)
-   * - Performance adjustments (efficiency factors for different conditions)
-   * - Analytical reporting of route segments and their data quality
-   *
-   * When examining a route, these flags provide important context about the
-   * reliability and characteristics of each segment.
-   */
-  enum DataMask {
-    /** Wind data originated from GRIB files. */
-    GRIB_WIND = 1,
-
-    /** Wind data originated from climatology data. */
-    CLIMATOLOGY_WIND = 2,
-
-    /**
-     * Wind data is from GRIB but is considered "data deficient".
-     * This typically means the data is from outside the requested time
-     * or location range but was used because better data was not available.
-     */
-    DATA_DEFICIENT_WIND = 4,
-
-    /** Current data originated from GRIB files. */
-    GRIB_CURRENT = 8,
-
-    /** Current data originated from climatology data. */
-    CLIMATOLOGY_CURRENT = 16,
-
-    /**
-     * Current data is from GRIB but is considered "data deficient".
-     * This typically means the data is from outside the requested time
-     * or location range but was used because better data was not available.
-     */
-    DATA_DEFICIENT_CURRENT = 32,
-
-    /**
-     * Indicates that this position occurs during nighttime.
-     * Used to apply nighttime efficiency factor and darker display colors.
-     */
-    NIGHT_TIME = 64
-  };
-  int data_mask;
-};
 
 /*
  * A RoutePoint that has a time associated with it, along with navigation and
@@ -356,676 +178,7 @@ public:
   double reflectivity;       //!< Reflectivity in dBZ.
 };
 
-class SkipPosition;
 class weather_routing_pi;
-
-enum PropagationError {
-  PROPAGATION_NO_ERROR = 0,
-  /** Position has already been propagated. */
-  PROPAGATION_ALREADY_PROPAGATED = 1,
-  /** Swell exceeds max value at position. */
-  PROPAGATION_EXCEEDED_MAX_SWELL = 2,
-  /** Position exceeds maximum latitude. */
-  PROPAGATION_EXCEEDED_MAX_LATITUDE = 3,
-  /** Wind data is missing or invalid. */
-  PROPAGATION_WIND_DATA_FAILED = 4,
-  /** Wind speed exceeds maximum value at position. */
-  PROPAGATION_EXCEEDED_MAX_WIND = 5,
-  /** Wind speed exceeds current speed at position. */
-  PROPAGATION_EXCEEDED_WIND_VS_CURRENT = 6,
-  /** Propagation angle is outside search limits. */
-  PROPAGATION_ANGLE_OUTSIDE_SEARCH_LIMITS = 7,
-  /** Propagation violates polar constraints. */
-  PROPAGATION_POLAR_CONSTRAINTS = 8,
-  /** Boat speed computation failed. */
-  PROPAGATION_BOAT_SPEED_COMPUTATION_FAILED = 9,
-  /** Apparent wind speed exceeds maximum value at position. */
-  PROPAGATION_EXCEEDED_APPARENT_WIND = 10,
-  /** Land intersection detected during propagation. */
-  PROPAGATION_LAND_INTERSECTION = 11,
-  /** Land safety margin exceeded during propagation. */
-  PROPAGATION_LAND_SAFETY_MARGIN = 12,
-  /** Boundary intersection detected during propagation. */
-  PROPAGATION_BOUNDARY_INTERSECTION = 13,
-  /** Cyclone track crossing detected during propagation. */
-  PROPAGATION_CYCLONE_TRACK_CROSSING = 14,
-  /** Propagation angle error. */
-  PROPAGATION_ANGLE_ERROR = 15
-};
-
-/**
- * Circular linked list node for positions which take equal time to reach.
- */
-class Position : public RoutePoint {
-public:
-  Position(double latitude, double longitude, Position* p = nullptr,
-           double pheading = NAN, double pbearing = NAN, int polar_idx = -1,
-           int tack_count = 0, int jibe_count = 0,
-           int sail_plan_change_count = 0, int data_mask = 0,
-           bool data_deficient = false);
-  Position(Position* p);
-
-  SkipPosition* BuildSkipList();
-
-  /**
-   * Propagates a position forward in time, generating new possible positions.
-   *
-   * This is a key routing function that expands the isochron by calculating all
-   * possible new positions reachable from the current position after the
-   * configured time step. It checks various constraints like wind, currents,
-   * land avoidance etc.
-   *
-   * @param routelist List to store the generated routes
-   * @param configuration Route configuration parameters
-   *
-   * @return true if propagation was successful and new routes were added,
-   *         false if no valid propagation was possible
-   */
-  bool Propagate(IsoRouteList& routelist, RouteMapConfiguration& configuration);
-
-  double Distance(Position* p);
-  // Return the number of times the sail configuration has changed.
-  int SailChanges();
-  double PropagateToEnd(RouteMapConfiguration& configuration, double& H,
-                        int& data_mask);
-
-  /** Helper method to get error as string. */
-  static wxString GetErrorText(PropagationError error);
-
-  /** Structure to track errors for specific angle attempts. */
-  struct AngleStatus {
-    double angle;  //!< The True Wind Angle that was attempted (degrees).
-    PropagationError error;  //!< Error code for this angle attempt.
-  };
-
-  /**
-   * Angle relative to true wind direction when traveling from parent to this
-   * position.
-   *
-   * This value represents the vessel's heading relative to the true wind when
-   * moving from the parent position to this position. It is measured in degrees
-   * where:
-   * - 0 degrees is directly upwind
-   * - 90 degrees is perpendicular to the wind (beam reach)
-   * - 180 degrees is directly downwind
-   *
-   * This angle is used to determine if tacking occurred (sign change in
-   * heading) and to apply tacking penalties. It's also used for optimizing sail
-   * configurations and evaluating polar performance.
-   */
-  double parent_heading;
-
-  /**
-   * Geographic bearing from parent position to this position.
-   *
-   * This value represents the compass bearing (in degrees) when traveling from
-   * the parent position to this position, where:
-   * - 0 degrees is North
-   * - 90 degrees is East
-   * - 180 degrees is South
-   * - 270 degrees is West
-   *
-   * This bearing is used to constrain the angular sector for the next
-   * propagation step based on the MaxSearchAngle configuration parameter.
-   */
-  double parent_bearing;
-
-  Position* parent;      /* previous position in time */
-  Position *prev, *next; /* doubly linked circular list of positions */
-
-  bool propagated;
-  bool drawn, copied;
-
-  /** Indicates why propagation failed. */
-  PropagationError propagation_error;
-
-private:
-  /** Reset error tracking information. */
-  void ResetErrorTracking();
-
-  /** Get error information for debugging. */
-  wxString GetErrorInfo() const;
-  /** Get detailed error information for debugging. */
-  wxString GetDetailedErrorInfo() const;
-
-  bool rk_step(double timeseconds, double cog, double dist, double twa,
-               RouteMapConfiguration& configuration, WR_GribRecordSet* grib,
-               const wxDateTime& time, int newpolar, double& rk_BG,
-               double& rk_dist, int& data_mask);
-};
-
-/**
- * A circular linked list that optimizes route path traversal by grouping
- * consecutive positions that share the same directional quadrant. This allows
- * the routing algorithm to quickly "skip" over large segments of consistent
- * direction, which improves performance when checking for intersections or
- * containment in routes with many points.
- *
- * Each SkipPosition points to:
- * 1. A Position object (a point in the route)
- * 2. The next SkipPosition in the list (which may skip over many intermediate
- * Positions)
- * 3. The previous SkipPosition in the list
- *
- * The quadrant value (0-3) represents the directional relationship between
- * consecutive positions, calculated based on their relative latitude and
- * longitude:
- * - Quadrant 0: Northwest movement (lat decreasing, lon decreasing)
- * - Quadrant 1: Northeast movement (lat decreasing, lon increasing)
- * - Quadrant 2: Southwest movement (lat increasing, lon decreasing)
- * - Quadrant 3: Southeast movement (lat increasing, lon increasing)
- */
-class SkipPosition {
-public:
-  /**
-   * Constructs a new SkipPosition object.
-   *
-   * @param p Pointer to a Position object that this SkipPosition will reference
-   * @param q Quadrant value (0-3) indicating the directional relationship
-   *          with the next position.
-   */
-  SkipPosition(Position* p, int q);
-
-  /**
-   * Removes this SkipPosition from the circular list.
-   *
-   * Connects the previous and next positions, maintaining the circular
-   * structure before deleting this instance.
-   */
-  void Remove();
-  /**
-   * Creates a deep copy of the entire skip list including referenced positions.
-   *
-   * This method recursively traverses the circular skip list, duplicating all
-   * SkipPosition objects along with their corresponding Position objects,
-   * maintaining the same structure and connectivity as the original list.
-   *
-   * @return Pointer to the first SkipPosition in the new copied list
-   */
-  SkipPosition* Copy();
-
-  Position* point;
-  SkipPosition *prev, *next;
-  /**
-   * Direction quadrant (0-3) indicating relationship between this point and
-   * the next.
-   */
-  int quadrant;
-};
-
-/**
- * Represents a closed loop of positions forming an isochrone boundary.
- *
- * An IsoRoute defines a closed contour of geographic positions that can be
- * reached within the same amount of time from a starting point. "Closed loop"
- * means the positions form a complete circuit where the last position connects
- * back to the first position, creating a continuous perimeter with no gaps or
- * endpoints.
- *
- * These boundaries (isochrones) expand outward as travel time increases,
- * creating concentric regions of accessibility from the starting point.
- *
- * The class supports:
- * - Normal regions (direction = 1) representing areas the vessel can reach.
- * - Inverted regions (direction = -1) representing "holes" or areas within a
- * normal region that cannot be reached (e.g., islands or other obstacles)
- * - Parent-child relationships for handling nested regions.
- *
- * IsoRoute provides methods for merging, normalizing (removing intersections),
- * checking containment, and propagating routes to create the next isochrone.
- */
-class IsoRoute {
-public:
-  /**
-   * Constructs an IsoRoute from a SkipPosition list.
-   *
-   * @param p Pointer to a SkipPosition that defines the start of the boundary
-   * @param dir Direction value: 1 for normal regions, -1 for inverted regions
-   */
-  IsoRoute(SkipPosition* p, int dir = 1);
-  /**
-   * Copy constructor that optionally sets a parent.
-   *
-   * Creates a deep copy of another IsoRoute, including all Position and
-   * SkipPosition objects, and optionally establishes a parent-child
-   * relationship.
-   *
-   * @param r Source IsoRoute to copy
-   * @param p Parent IsoRoute (nullptr if no parent)
-   */
-  IsoRoute(IsoRoute* r, IsoRoute* p = nullptr);
-  ~IsoRoute();
-
-  /**
-   * Outputs route information for debugging.
-   *
-   * Prints all positions in the route to standard output for debugging
-   * purposes.
-   */
-  void Print();
-  /**
-   * Outputs skip list information for debugging.
-   *
-   * Prints all skip positions in the route to standard output for debugging
-   * purposes.
-   */
-  void PrintSkip();
-
-  /**
-   * Adjusts skippoints to start at the minimum latitude.
-   *
-   * Ensures the skippoints list starts at the position with minimum latitude,
-   * which helps identify the outer boundary of the route for intersection
-   * testing.
-   */
-  void MinimizeLat();
-  /**
-   * Counts intersections between a position and this route.
-   *
-   * Determines how many times a ray from the given position to infinity
-   * crosses the route boundary. Used to determine if a position is inside
-   * or outside the route.
-   *
-   * @param pos Position to test
-   * @return Number of intersections, or -1 if the test is inconclusive
-   */
-  int IntersectionCount(Position& pos);
-  /**
-   * Tests if a position is contained within this route.
-   *
-   * Uses ray casting algorithm to determine if a position is inside this route,
-   * optionally testing children routes as well.
-   *
-   * @param pos Position to test
-   * @param test_children Whether to test child routes as well
-   * @return 1 if position is inside, 0 if outside, -1 if test is inconclusive
-   */
-  int Contains(Position& pos, bool test_children);
-
-  /**
-   * Tests if this route completely contains another route.
-   *
-   * Checks if every position in route r is contained within this route.
-   *
-   * @param r The route to test for containment
-   * @return true if r is completely contained within this route
-   */
-  bool CompletelyContained(IsoRoute* r);
-  /**
-   * Tests if this route contains another route.
-   *
-   * Tests if the first position of route r is contained within this route.
-   * If that test is inconclusive, tries other positions.
-   *
-   * @param r The route to test for containment
-   * @return true if r is contained within this route
-   */
-  bool ContainsRoute(IsoRoute* r);
-
-  /**
-   * Simplifies the route by removing points that are very close together.
-   *
-   * Removes positions that are within a small epsilon distance of each other
-   * to improve computational efficiency without significantly changing the
-   * route.
-   */
-  void ReduceClosePoints();
-  //    bool ApplyCurrents(GribRecordSet *grib, wxDateTime time,
-  //    RouteMapConfiguration &configuration);
-  /**
-   * Finds the bounding box of this route.
-   *
-   * Calculates the minimum and maximum latitude and longitude values
-   * that encompass the entire route.
-   *
-   * @param bounds[4] Array to store results: [min_lon, max_lon, min_lat,
-   * max_lat]
-   */
-  void FindIsoRouteBounds(double bounds[4]);
-
-  /**
-   * Removes a position from the route.
-   *
-   * Removes and deletes a position from the route, updating the skip list
-   * if necessary.
-   *
-   * @param s The SkipPosition preceding the position to remove
-   * @param p The Position to remove
-   */
-  void RemovePosition(SkipPosition* s, Position* p);
-  /**
-   * Finds the closest position in the route to a given lat/lon.
-   *
-   * Searches the route and its children for the position closest to
-   * the specified coordinates.
-   *
-   * @param lat Latitude to find closest position to
-   * @param lon Longitude to find closest position to
-   * @param dist Pointer to store the squared distance (optional)
-   * @return Pointer to the closest position
-   */
-  Position* ClosestPosition(double lat, double lon, double* dist = 0);
-  /**
-   * Attempts to propagate from all positions in this route.
-   *
-   * Creates new potential routes by propagating from each position in this
-   * route according to the specified configuration.
-   *
-   * @param routelist List to store new routes generated by propagation
-   * @param configuration Route configuration parameters
-   * @return true if at least one successful propagations occurred.
-   */
-  bool Propagate(IsoRouteList& routelist, RouteMapConfiguration& configuration);
-  /**
-   * Propagates directly to the configured end position.
-   *
-   * Attempts to find the best path from this route directly to the
-   * destination point specified in the configuration.
-   *
-   * @param configuration Route configuration parameters
-   * @param mindt [out] Minimum time to reach destination
-   * @param endp [out] Position from which the end is reached
-   * @param minH [out] Final heading at destination
-   * @param mintacked [out] Whether tacking occurred on final approach
-   * @param minjibed [out] Whether jibing occurred on final approach
-   * @param minsail_plan_changed [out] Whether sailplan changed on final
-   * approach
-   * @param mindata_mask [out] Data source mask for the final approach
-   */
-  void PropagateToEnd(RouteMapConfiguration& configuration, double& mindt,
-                      Position*& endp, double& minH, bool& mintacked,
-                      bool& minjibed, bool& minsail_plan_changed,
-                      int& mindata_mask);
-
-  /**
-   * Counts the number of skip positions in this route.
-   *
-   * @return Number of skip positions
-   */
-  int SkipCount();
-  /**
-   * Counts the total number of positions in this route.
-   *
-   * @return Number of positions
-   */
-  int Count();
-  /**
-   * Updates various statistics about this route and its children.
-   *
-   * Calculates totals for routes, inverted routes, skip positions,
-   * and positions across this route and all its children.
-   *
-   * @param routes [in/out] Counter for number of routes
-   * @param invroutes [in/out] Counter for number of inverted routes
-   * @param skippositions [in/out] Counter for number of skip positions
-   * @param positions [in/out] Counter for number of positions
-   */
-  void UpdateStatistics(int& routes, int& invroutes, int& skippositions,
-                        int& positions);
-  /**
-   * Resets the drawn flag on all positions in this route and its children.
-   *
-   * Marks all positions as not drawn, used before rendering the route.
-   */
-  void ResetDrawnFlag();
-
-  /**
-   * Skip list of positions for efficient traversal.
-   *
-   * The skippoints list provides optimized access to the positions in this
-   * route, allowing rapid traversal of segments that maintain the same
-   * directional quadrant.
-   */
-
-  SkipPosition* skippoints;
-
-  /**
-   * Indicates whether this is a normal (1) or inverted (-1) region.
-   *
-   * Direction = 1: Normal region the vessel can reach.
-   * Direction = -1: Inverted region (hole) representing areas that cannot be
-   * reached.
-   */
-  int direction;
-
-  /**
-   * Pointer to the parent route if this is a child (inner region).
-   *
-   * nullptr if this is a top-level route, otherwise points to the outer region
-   * that contains this route.
-   */
-  IsoRoute* parent;
-  /**
-   * List of child routes representing inner inverted regions.
-   *
-   * Contains routes representing areas inside this route that cannot be
-   * reached (e.g., islands, obstacles, or other exclusion zones).
-   */
-  IsoRouteList children;
-};
-
-// -----------------
-class WR_GribRecordSet {
-public:
-  WR_GribRecordSet(unsigned int id) : m_Reference_Time(-1), m_ID(id) {
-    for (int i = 0; i < Idx_COUNT; i++) {
-      m_GribRecordPtrArray[i] = 0;
-      m_GribRecordUnref[i] = false;
-    }
-  }
-
-  virtual ~WR_GribRecordSet() { RemoveGribRecords(); }
-
-  /* copy and paste by plugins, keep functions in header */
-  void SetUnRefGribRecord(int i, GribRecord* pGR) {
-    assert(i >= 0 && i < Idx_COUNT);
-    if (m_GribRecordUnref[i] == true) {
-      delete m_GribRecordPtrArray[i];
-    }
-    m_GribRecordPtrArray[i] = pGR;
-    m_GribRecordUnref[i] = true;
-  }
-
-  void RemoveGribRecords() {
-    for (int i = 0; i < Idx_COUNT; i++) {
-      if (m_GribRecordUnref[i] == true) {
-        delete m_GribRecordPtrArray[i];
-      }
-    }
-  }
-
-  time_t m_Reference_Time;
-  unsigned int m_ID;
-
-  GribRecord* m_GribRecordPtrArray[Idx_COUNT];
-
-private:
-  // grib records files are stored and owned by reader mapGribRecords
-  // interpolated grib are not, keep track of them
-  bool m_GribRecordUnref[Idx_COUNT];
-};
-
-// ------
-class Shared_GribRecordSetData : public wxRefCounter {
-public:
-  Shared_GribRecordSetData(WR_GribRecordSet* gribset = 0)
-      : m_GribRecordSet(gribset) {}
-  Shared_GribRecordSetData(const Shared_GribRecordSetData& data)
-      : m_GribRecordSet(data.m_GribRecordSet) {}
-
-  void SetGribRecordSet(WR_GribRecordSet* gribset) {
-    m_GribRecordSet = gribset;
-  }
-  WR_GribRecordSet* GetGribRecordSet() const { return m_GribRecordSet; }
-
-  ~Shared_GribRecordSetData();
-
-protected:
-  WR_GribRecordSet* m_GribRecordSet;
-};
-
-// ------
-class Shared_GribRecordSet : public wxTrackable {
-public:
-  // initializes this, assigning to the
-  // internal data pointer a new instance of Shared_GribRecordSetData
-  Shared_GribRecordSet(WR_GribRecordSet* ptr = 0)
-      : m_data(new Shared_GribRecordSetData(ptr)) {}
-  Shared_GribRecordSet& operator=(const Shared_GribRecordSet& tocopy) {
-    // shallow copy: this is just a fast copy of pointers; the real
-    // memory-consuming data which typically is stored inside
-    m_data = tocopy.m_data;
-    return *this;
-  }
-
-  void SetGribRecordSet(WR_GribRecordSet* ptr) {
-    // make sure changes to this class do not affect other instances
-    // currently sharing our same refcounted data:
-    UnShare();
-    m_data->SetGribRecordSet(ptr);
-  }
-
-  WR_GribRecordSet* GetGribRecordSet() const {
-    return m_data->GetGribRecordSet();
-  }
-
-  bool operator==(const Shared_GribRecordSet& other) const {
-    if (m_data.get() == other.m_data.get())
-      return true;  // this instance and the 'other' one share the same data...
-    return (m_data->GetGribRecordSet() == other.m_data->GetGribRecordSet());
-  }
-
-  wxObjectDataPtr<Shared_GribRecordSetData> m_data;
-
-protected:
-  void UnShare() {
-    if (m_data->GetRefCount() == 1) return;
-    m_data.reset(new Shared_GribRecordSetData(*m_data));
-  }
-};
-
-/**
- * Manages a collection of IsoRoute objects that represent equal-time boundaries
- * from a starting point.
- *
- * While an IsoRoute represents a single closed boundary, an IsoChron
- * encapsulates a complete set of multiple IsoRoute objects that together define
- * all regions reachable within the same amount of time.
- *
- * An IsoChron contains:
- * - A list of IsoRoute objects (which may include normal regions and their
- * inverted "hole" regions)
- * - The specific time value this isochrone represents (when these points can be
- * reached)
- * - The time delta from the previous isochrone
- * - Associated weather data (GRIB) for this time period
- *
- * The IsoChron class forms the backbone of the routing algorithm, with
- * successive IsoChron objects representing expanding time boundaries as the
- * vessel travels further from the origin. Each IsoChron builds upon the
- * previous one by propagating routes outward based on vessel capabilities and
- * environmental conditions.
- */
-class IsoChron {
-public:
-  /**
-   * Constructs an IsoChron from a list of routes and associated time and
-   * weather data.
-   *
-   * @param r List of IsoRoute objects forming this isochrone
-   * @param t Time when the vessel would reach positions on this isochrone
-   * @param d Time increment (in seconds) from previous isochrone
-   * @param g Shared GRIB weather data associated with this time period
-   * @param grib_is_data_deficient Flag indicating if GRIB data has limitations
-   */
-  IsoChron(IsoRouteList r, wxDateTime t, double d, Shared_GribRecordSet& g,
-           bool grib_is_data_deficient);
-  ~IsoChron();
-
-  /**
-   * Propagates all routes in this isochron to create the next isochrone.
-   *
-   * From each position on each route in the current isochrone, calculates all
-   * possible new positions the vessel could reach in the next time increment,
-   * creating a new expanded set of routes.
-   *
-   * @param routelist Output list to store the newly generated routes
-   * @param configuration Route configuration parameters controlling propagation
-   */
-  void PropagateIntoList(IsoRouteList& routelist,
-                         RouteMapConfiguration& configuration);
-  /**
-   * Tests if a position is contained within any route in this isochrone.
-   *
-   * @param p Position to test
-   * @return true if the position is inside any route in this isochrone
-   */
-  bool Contains(Position& p);
-  /**
-   * Tests if a geographic coordinate is contained within any route in this
-   * isochrone.
-   *
-   * @param lat Latitude to test
-   * @param lon Longitude to test
-   * @return true if the coordinate is inside any route in this isochrone
-   */
-  bool Contains(double lat, double lon);
-  /**
-   * Finds the closest position in any route to the given coordinates.
-   *
-   * @param lat Latitude to find closest position to
-   * @param lon Longitude to find closest position to
-   * @param t Optional pointer to store the time at the closest position
-   * @param dist Optional pointer to store the distance to the closest position
-   * @return Pointer to the closest position
-   */
-  Position* ClosestPosition(double lat, double lon, wxDateTime* t = 0,
-                            double* dist = 0);
-  void ResetDrawnFlag();
-
-  /**
-   * List of IsoRoute objects that together form this isochrone.
-   *
-   * Each IsoRoute represents a distinct region reachable within the same time,
-   * and collectively they define the complete isochrone boundary.
-   */
-  IsoRouteList routes;
-  /**
-   * The time at which all positions on this isochrone can be reached.
-   *
-   * All routes in this isochrone have equal travel time from the starting
-   * point.
-   */
-  wxDateTime time;
-  /**
-   * The time increment (in seconds) from the previous isochrone to this one.
-   *
-   * Represents how much more time is needed to reach this isochrone compared
-   * to the previous one.
-   */
-  double delta;
-  /**
-   * Shared reference to GRIB weather data for this time period.
-   *
-   * Contains wind, current, and other weather information used for
-   * calculations.
-   */
-  Shared_GribRecordSet m_SharedGrib;
-  /**
-   * Pointer to the GRIB record set for this time period.
-   *
-   * Direct access to weather data used in propagation calculations.
-   */
-  WR_GribRecordSet* m_Grib;
-  /**
-   * Flag indicating if the GRIB data has limitations or deficiencies.
-   *
-   * When true, weather data may be incomplete or extrapolated.
-   */
-  bool m_Grib_is_data_deficient;
-};
-
-typedef std::list<IsoChron*> IsoChronList;
 
 /**
  * Represents a named geographic position that can be used as a start or end
