+++ src/ConstraintChecker.cpp
@@ -1,0 +1,35 @@
+/***************************************************************************
+ *   Copyright (C) 2015 by OpenCPN development team                        *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 3 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program; if not, write to the                         *
+ *   Free Software Foundation, Inc.,                                       *
+ *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301,  USA.         *
+ ***************************************************************************/
+
+#include <wx/wx.h>
+
+#include "ConstraintChecker.h"
+#include "WeatherDataProvider.h"
+#include "RouteMap.h"
+
+bool ConstraintChecker::CheckSwellConstraint(
+    double lat, double lon, RouteMapConfiguration& configuration,
+    PropagationError& error_code) {
+  double swell = WeatherDataProvider::GetSwell(configuration, lat, lon);
+  if (swell > configuration.MaxSwellMeters) {
+    error_code = PROPAGATION_EXCEEDED_MAX_SWELL;
+    return false;
+  }
+  return true;
+}
